fix: build.sh script for MSYS2/MINGW64
When running `build/build.sh` (which is called from `Makefile`),
the script fails as Cygwin is not a valid `OSTYPE` in the `checkOS`
function. This PR adds Cygwin as one, recognizing it as Windows.

Additionally, the `checkOS` function also properly receives the first
command line argument if it was not able to detect an `OSTYPE`, instead of
being left empty if the argument was present.