fix: Remove bad common commands.
They cause far more problems than they solve, and now with the existence of the InputTime trigger you can detect a button press without a command, solving the problems that these originally solved.

I presume the blank "recovery" command is there so that the common states referring to it don't crash the game if the character doesn't have the command, but is there any reason the game should crash from this to begin with?