feat: X/Y Truncate setting
This commit adds a new video setting called X/Y Truncate or X/Y Rounding.
When enabled, it will forcibly clamp the X and Y coords of all 2D assets
to integer numbers just before they get rendered to the screen.
(e.g. [152.49, 202.17] becomes [152.00, 202.00])

This helps to deal with a sort of "blurring" artifact that occurs on low-res
characters when playing with MSAA on, which is particularly noticeable
when doing short dashes by characters such as Kung Fu Man.

The reason I made this an option is that it may also cause off-by-one-pixel
effects in certain cases (namely screenpacks), and thus it can't just be
something that's done all the time. Still, it's enough of an improvement
that I opted to enable it by default.