Add ProjClsnOverlap trigger
・Add ProjClsnOverlap trigger
The format:
ProjClsnOverlap(index, Target_PlayerID, box_type)
This index is equivalent to the index when -1 is specified for the ID in ProjVar.

・Moved xAccel from common1.cns.zss to system.zss
This makes xAccel enable even for characters that do not use common1.cns.zss.
Additionally, zAccel is now enable.

・Fixed a bug where changes to char slots were not reflected during multiplayer

・Fixed Trans Sctrl so that specifying default results in "do nothing"

・Relaxed the conditions that cause a crash when using both old and new syntax together in Hitby
This because a few of the chars I created long ago was crashing.
Additionally, I confirmed that several other chars were also crashing.