Fix PlaySnd volume
・WINMUGEN has a bug where the volume parameter is disabled when loop is specified. (volumescale is not affected by this).

・Change the range of values accepted by PlaySnd volume parameter from [-256 to 256] to [-512 to 512]
This is just temporarily widening the range of values that volume accepts, and it does not accurately reproduce WinMUGEN's behavior. In WINMUGEN , the playsnd volume parameter is affected by const [Data]volume, but that logic seems different from the current IKEMEN GO implementation. but I'm not sure if it needs to be perfectly replicated.

・Fixes the pos of Afterimage for Explod with space=screen.