fix: persistent value when changestate during hitpause
・Fix persistent value when changestate during hitpause

1.Carrying Over a Counter to a persistent = 0 SCTRL:
When a SCTRL with persistent = 0 inherits a persistent counter from a previous state that was still counting down (value was > 0), the new SCTRL will now be treated as if it has already been executed. Its counter is immediately set to a locked state (math.MaxInt32), preventing it from running until the character's persistent counters are next reset. If the inherited counter was already ≤ 0 (ready to execute), the SCTRL remains executable.

2.Carrying Over an Executed persistent = 0 Counter to a persistent > 0 SCTRL:
Conversely, when an SCTRL with a repeating persistent value (e.g., persistent = 32) inherits a locked counter (math.MaxInt32) from a persistent = 0 SCTRL that had already run, the counter is now reset to the new SCTRL's persistent value. This unlocks the SCTRL.

・Fix Hitdef keepstate
Damage handling for HitOverride
Moved the process of pausing the opponent to system.zss